fix(ci): use target branch instead of camunda version (#878)
* fix(ci): use target branch instead of camunda version

This PR switch usage of the camunda version in favor of the target
branch, it reduces (but do not limit in case of a breaking change
between the PR and the target branch), the number of incompatibilities.
Especially when main is targeting a camunda version that is also
referenced in an other branch.

* fix